more fixes for Dome version

+++ SpellsFromTheWest/FuLongTan/Tuifa3.cs
@@ -72,7 +72,7 @@ namespace GameData.Domains.SpecialEffect.MoreFactionCombatSkills.FuLongTan
             }
         }
 
-        private void OnAddInjury(DataContext context, CombatCharacter character, sbyte bodyPart, bool isInner, sbyte value, bool changeToOld)
+        private void OnAddInjury(DataContext context, CombatCharacter character, sbyte bodyPart, bool isInner, int value, bool changeToOld)
         {
             if (_affecting || _delaying)
             {
