+++ SpellsFromTheWest/FuLongTan/Tuifa8.cs
@@ -51,7 +51,7 @@ namespace GameData.Domains.SpecialEffect.MoreFactionCombatSkills.FuLongTan
             Events.UnRegisterHandler_CastSkillEnd(OnCastSkillEnd);
         }
 
-        private void OnAddInjury(DataContext context, CombatCharacter character, sbyte bodyPart, bool isInner, sbyte value, bool changeToOld)
+        private void OnAddInjury(DataContext context, CombatCharacter character, sbyte bodyPart, bool isInner, int value, bool changeToOld)
         {
             if (_affecting || _delaying)
             {
